\name{thanojRandomLetter-package}
\alias{thanojRandomLetter-package}
\alias{thanojRandomLetter}
\docType{package}
\title{
\packageTitle{thanojRandomLetter}
}
\description{
\packageDescription{thanojRandomLetter}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{thanojRandomLetter}
\packageIndices{thanojRandomLetter}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{thanojRandomLetter}

Maintainer: \packageMaintainer{thanojRandomLetter}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
